Move scenario.hpp/.cpp to boe library
