+++ include/boe/scenario.hpp
